# Add another solution (a bit faster)