> Performance (SubstringPairsDigraphsLookup):
>   1000 runs
>   17128ms total run time
>   Average time: 0.017128
>   | Min | 25% | 50% | 75% | 95% | Max |
>   | 016 | 017 | 017 | 017 | 019 | 022 |

> Performance (BitShiftingBinarySearchDigraphsLookup):
>   1000 runs
>   1154ms total run time
>   Average time: 0.001154
>   | Min | 25% | 50% | 75% | 95% | Max |
>   | 001 | 001 | 001 | 001 | 002 | 002 |

> Performance (MemoryArraySearchDigraphsLookup):
>   1000 runs
>   59ms total run time
>   Average time: 5.9E-05
>   | Min | 25% | 50% | 75% | 95% | Max |
>   | 000 | 000 | 000 | 000 | 001 | 001 |

> Performance (MemoryArraySearchDigraphsLookupV2):
>   1000 runs
>   70ms total run time
>   Average time: 7E-05
>   | Min | 25% | 50% | 75% | 95% | Max |
>   | 000 | 000 | 000 | 000 | 001 | 001 |