@christianrondeau 2018-05-16
Very good! This is a clear winner :) Good idea, replacing operations by a larger memory space.

Note that one of the test was broken :P (But it's a detail).

Suggestions:

The first loop that sets results to 0 is not useful; an array is always initalized to zero:

```
			for (var i = 0; i < _result.Length; i++)
			{
				_result[i] = 0;
			}
```

The initialization of the array in the constructor is a little bit cheating, since building the object is not part of the perf test ;)

But otherwise, bravo!